SshProxy: pass some more functions to the client seat.
Now that we're actually using it for messages, we also need to pass
its interactivity setting through to the subsidiary SSH backend,
because otherwise that won't know whether to display particular
messages. Same goes for constructing a StripCtrl for SSH auth banners
(which wants to be done the same way between primary and proxy SSH
connections), and so on.