Tidy up the comments in PlugVtable.
It always confused me that each comment was _after_ the function
prototype it described, instead of before, which is my usual idiom.
Reordered everything, and added a blank line between each
(comment,function) pair to make it clear what goes with what.

While I'm at it, rewrote some of the comments for clarity and whole
sentences.